Imported ARDUnity from the Asset Store.
In order to get the .cpp and .h files for the arduino, I had to import the ARDUnity package from the assets store.

TODO: I'm not really sure if I shoud be versioning imported assets in general. It doesn't look like there is anything OS specific in here, but I'm not sure what the normal procedure is.